Move AppIssue and IssueSeverity to error panel
